[TDL-192] Rename email forwarders per ADR #12
