+++ libs/common/src/email-forwarders/anon-addy-forwarder.ts
@@ -1,28 +1,29 @@
 import { ApiService } from "../abstractions/api.service";
 
 import { Forwarder } from "./forwarder";
-import { ForwarderOptions } from "./forwarderOptions";
+import { ForwarderOptions } from "./forwarder-options";
 
-export class SimpleLoginForwarder implements Forwarder {
+export class AnonAddyForwarder implements Forwarder {
   async generate(apiService: ApiService, options: ForwarderOptions): Promise<string> {
     if (options.apiKey == null || options.apiKey === "") {
-      throw "Invalid SimpleLogin API key.";
+      throw "Invalid AnonAddy API token.";
+    }
+    if (options.anonaddy?.domain == null || options.anonaddy.domain === "") {
+      throw "Invalid AnonAddy domain.";
     }
     const requestInit: RequestInit = {
       redirect: "manual",
       cache: "no-store",
       method: "POST",
       headers: new Headers({
-        Authentication: options.apiKey,
+        Authorization: "Bearer " + options.apiKey,
         "Content-Type": "application/json",
       }),
     };
-    let url = "https://app.simplelogin.io/api/alias/random/new";
-    if (options.website != null) {
-      url += "?hostname=" + options.website;
-    }
+    const url = "https://app.anonaddy.com/api/v1/aliases";
     requestInit.body = JSON.stringify({
-      note:
+      domain: options.anonaddy.domain,
+      description:
         (options.website != null ? "Website: " + options.website + ". " : "") +
         "Generated by Bitwarden.",
     });
@@ -30,19 +31,11 @@ export class SimpleLoginForwarder implements Forwarder {
     const response = await apiService.nativeFetch(request);
     if (response.status === 200 || response.status === 201) {
       const json = await response.json();
-      return json.alias;
+      return json?.data?.email;
     }
     if (response.status === 401) {
-      throw "Invalid SimpleLogin API key.";
+      throw "Invalid AnonAddy API token.";
     }
-    try {
-      const json = await response.json();
-      if (json?.error != null) {
-        throw "SimpleLogin error:" + json.error;
-      }
-    } catch {
-      // Do nothing...
-    }
-    throw "Unknown SimpleLogin error occurred.";
+    throw "Unknown AnonAddy error occurred.";
   }
 }
