+++ libs/common/src/email-forwarders/forwarder.ts
@@ -1,6 +1,6 @@
 import { ApiService } from "../abstractions/api.service";
 
-import { ForwarderOptions } from "./forwarderOptions";
+import { ForwarderOptions } from "./forwarder-options";
 
 export interface Forwarder {
   generate(apiService: ApiService, options: ForwarderOptions): Promise<string>;
