+++ libs/common/src/email-forwarders/simple-login-forwarder.ts
@@ -1,29 +1,28 @@
 import { ApiService } from "../abstractions/api.service";
 
 import { Forwarder } from "./forwarder";
-import { ForwarderOptions } from "./forwarderOptions";
+import { ForwarderOptions } from "./forwarder-options";
 
-export class AnonAddyForwarder implements Forwarder {
+export class SimpleLoginForwarder implements Forwarder {
   async generate(apiService: ApiService, options: ForwarderOptions): Promise<string> {
     if (options.apiKey == null || options.apiKey === "") {
-      throw "Invalid AnonAddy API token.";
-    }
-    if (options.anonaddy?.domain == null || options.anonaddy.domain === "") {
-      throw "Invalid AnonAddy domain.";
+      throw "Invalid SimpleLogin API key.";
     }
     const requestInit: RequestInit = {
       redirect: "manual",
       cache: "no-store",
       method: "POST",
       headers: new Headers({
-        Authorization: "Bearer " + options.apiKey,
+        Authentication: options.apiKey,
         "Content-Type": "application/json",
       }),
     };
-    const url = "https://app.anonaddy.com/api/v1/aliases";
+    let url = "https://app.simplelogin.io/api/alias/random/new";
+    if (options.website != null) {
+      url += "?hostname=" + options.website;
+    }
     requestInit.body = JSON.stringify({
-      domain: options.anonaddy.domain,
-      description:
+      note:
         (options.website != null ? "Website: " + options.website + ". " : "") +
         "Generated by Bitwarden.",
     });
@@ -31,11 +30,19 @@ export class AnonAddyForwarder implements Forwarder {
     const response = await apiService.nativeFetch(request);
     if (response.status === 200 || response.status === 201) {
       const json = await response.json();
-      return json?.data?.email;
+      return json.alias;
     }
     if (response.status === 401) {
-      throw "Invalid AnonAddy API token.";
+      throw "Invalid SimpleLogin API key.";
     }
-    throw "Unknown AnonAddy error occurred.";
+    try {
+      const json = await response.json();
+      if (json?.error != null) {
+        throw "SimpleLogin error:" + json.error;
+      }
+    } catch {
+      // Do nothing...
+    }
+    throw "Unknown SimpleLogin error occurred.";
   }
 }
